Add files via upload
This is a PowerShell script that I put together for a project at work. I had a lot of help from ChatGPT with it. My company released a new Group Policy forcing everyone to use OneDrive to backup their files. After this happened, a few users had issues where a lot of their files would duplicate themselves and add "- Copy." I made a tool to help with that. This tool looks at a user specified folder in Windows and will scan that folder using the Sha256 hash and if there are any duplicates, the ones with '- Copy' in their file name are moved to a new folder that I put in the C:\ directory.